+++ src/SWP/Bundle/CoreBundle/Context/ScopeContextInterface.php
@@ -21,4 +21,6 @@ interface ScopeContextInterface extends BaseScopeContextInterface
     const SCOPE_ORGANIZATION = 'organization';
 
     const SCOPE_TENANT = 'tenant';
+
+    const SCOPE_THEME = 'theme';
 }