+++ src/SWP/Bundle/SettingsBundle/Provider/ConfigurationSettingsProvider.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Superdesk Web Publisher Settings Bundle.
+ *
+ * Copyright 2018 Sourcefabric z.ú. and contributors.
+ *
+ * For the full copyright and license information, please see the
+ * AUTHORS and LICENSE files distributed with this source code.
+ *
+ * @copyright 2018 Sourcefabric z.ú
+ * @license http://www.superdesk.org/license
+ */
+
+namespace SWP\Bundle\SettingsBundle\Provider;
+
+final class ConfigurationSettingsProvider implements SettingsProviderInterface
+{
+    /**
+     * @var array
+     */
+    private $settingsConfiguration = [];
+
+    /**
+     * ConfigurationSettingsProvider constructor.
+     *
+     * @param array $settingsConfiguration
+     */
+    public function __construct(array $settingsConfiguration = [])
+    {
+        $this->settingsConfiguration = $settingsConfiguration;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSettings(): array
+    {
+        return $this->settingsConfiguration;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supports(): bool
+    {
+        return true;
+    }
+}