+++ src/SWP/Bundle/SettingsBundle/Provider/SettingsProviderInterface.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Superdesk Web Publisher Settings Bundle.
+ *
+ * Copyright 2018 Sourcefabric z.ú. and contributors.
+ *
+ * For the full copyright and license information, please see the
+ * AUTHORS and LICENSE files distributed with this source code.
+ *
+ * @copyright 2018 Sourcefabric z.ú
+ * @license http://www.superdesk.org/license
+ */
+
+namespace SWP\Bundle\SettingsBundle\Provider;
+
+interface SettingsProviderInterface
+{
+    /**
+     * @return array
+     */
+    public function getSettings(): array;
+
+    /**
+     * @return bool
+     */
+    public function supports(): bool;
+}